Preserve config registry across plugin reinstall

+++ uninstall.php
@@ -36,4 +36,6 @@ delete_option( 'xpressui_workflow_manifest_registry' );
 delete_option( 'xpressui_bundled_workflows_installed' );
 delete_option( 'xpressui_bundled_workflows_version' );
 delete_option( 'xpressui_user_deleted_workflows' );
-delete_option( 'xpressui_project_config_registry' );
+if ( $xpressui_delete_submissions ) {
+	delete_option( 'xpressui_project_config_registry' );
+}
